# Remove dependency on vLLM ScalarType for mixed_precision_gemm interface