### Description

This PR removes the dependency on vLLM's `ScalarType` class for the public interface of `mixed_precision_gemm`. This will make it easier to provide support for Conch's mixed_precision_gemm inside of vLLM.

### Testing

Please select all that apply.

- [x] Existing unit tests
- [ ] Unit tests added by this PR
- [ ] Other (please explain)
- [ ] This PR is not tested

#### Test instructions

```
pytest tests/mixed_precision_gemm_test.py
```

```
# On H100, add CONCH_ENABLE_VLLM=1 and --enable-machete
python benchmarks/mixed_precision_gemm_benchmark.py
```

**A10**

```
48 passed in 4.07s
```

```
Results matched :)
Parameters: {'m_dim': 4096, 'k_dim': 8192, 'n_dim': 4096, 'input_dtype': 'fp16', 'weight_dtype': 'uint4b8'}
Triton: num_iterations=100, min=10.580 ms, max=16.913 ms, mean=11.023 ms, median=10.757 ms
```

**H100**

```
48 passed in 8.03s
```

```
INFO 06-09 17:19:37 [importing.py:53] Triton module has been replaced with a placeholder.
INFO 06-09 17:19:38 [__init__.py:239] Automatically detected platform cuda.
Results matched :)
Parameters: {'m_dim': 4096, 'k_dim': 8192, 'n_dim': 4096, 'input_dtype': 'fp16', 'weight_dtype': 'uint4b8'}
Triton: num_iterations=100, min=1.756 ms, max=1.944 ms, mean=1.794 ms, median=1.783 ms
Baseline: num_iterations=100, min=0.438 ms, max=0.474 ms, mean=0.450 ms, median=0.443 ms
```

**MI300X**

```
48 passed in 16.07s
```

```
Results matched :)
Parameters: {'m_dim': 4096, 'k_dim': 8192, 'n_dim': 4096, 'input_dtype': 'fp16', 'weight_dtype': 'uint4b8'}
Triton: num_iterations=100, min=1.219 ms, max=1.424 ms, mean=1.270 ms, median=1.258 ms
```

#### Platforms

Please select all hardware platforms that this PR was tested on.

- [x] Nvidia GPU
- [x] AMD GPU
- [ ] Other (please explain)